refactor backend tests to reduce duplication
